test(init): isolate init regressions
